requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
idna==3.3
packaging==21.3
pyparsing==3.0.7
python-gitlab==3.3.0
PyYAML==6.0
requests==2.27.1
requests-toolbelt==0.9.1
urllib3==1.26.9
